Pin session interfaces
This patch add the session interfaces for controlling and observing GPIO
pins. The "Pin_state" session allows a client to request a pin state
whereas the "Pin_contol" session can be used to drive an output pin.

The session interfaces are accompanied by utilities for implementing
them. In particular, the component.h files contain ready-to-use
'Session_component' implementations, and 'os/pin_driver.h' contains
a generic 'Root' component implementation along with interfaces and
utilities for driver implementations.

This patch is designated to become a part of Genode repository.