pio driver: allow time multiplexing of pin in/out
This patch enhances the PIO pin driver with the ability to operate one
GPIO pin sequentially as input or output. This is required for devices
like the Goodix touchscreen controllers that reuse an input signal line
(IRQ line) for configuration purposes during device reset.

The time multiplexing works by declaring the same pin as both <in> and
and <out> node without a default attribute. The pin is switched into
output mode while a pin-control session is associated with the pin.
The pin can be repurposed for an IRQ session by closing the
pin-control session.