Unify common lx_emul parts across drivers
By moving dummies that are same for each driver to a central place
(lx_emul/a64/common_dummies.cc), this patch reduces code repetition and
eases the porting of further drivers.

The __cond_resched function that is moved to lx_emul/a64/sched.c may be
a candidate for dde_linux.