Initial phone version of Sculpt OS
This patch introduces the phone_manager as a replacement of the
sculpt_manager for the phone use case. This initial version is able to
control the power state of the modem, interact with the SIM card,
initiate phone calls via a dial pad, and pick up inbound calls.
At the current time, the speaker of the PinePhone is unconditionally
enabled.

Note that this intermediate version does not offer all features of the
regular Sculpt OS yet. E.g., switching back and forth between the
administrative GUI and the runtime, or the dynamic deployment of
packages are not yet supported.

The new phone variant can be built via the following command:

  make run/sculpt KERNEL=hw BOARD=pinephone SCULPT=phone

A mockup version of the phone UI can be started on Linux via:

  make run/sculpt_test KERNEL=linux BOARD=linux SCULPT=phone

Issue #5