at_protocol: application of QCFG settings
QCFG settings - such as the usbnet mode of operation - happen to be
stored persistently on the modem. When starting the modem driver, it
cannot assume the current settings to be compatible with the driver's
assumptions (e.g., that the modem appears as USB ECM device).
Unfortunately, the intuitive approach of unconditionally setting the
driver's desired values at start time is not practical become certain
settings ("usbnet") require a modem reboot in order to become effective.

This patch extends the AT protocol implementation with the support for
the validation and - if needed - adjustment of QCFG settings. At start,
the driver queries the current settings and compares them to the desired
values. Mismatching values are re-assigned to the desired values. Once
this QCFG value synchronization is complete, a modem reboot is scheduled
if any value was changed.

While testing the mechanism on the PinePhone, it turned out that certain
QCFG attributes - "usbnet" being an example - cannot be reliably queried
at the modem's boot time. The resulting "ERROR" must be handled by
retrying a operation at a later time. Hence, this patch generalizes the
former sim-busy handling to capture this corner case.

As the QCFG synchronization cannot be performed too early, this patch
adjusts the priorization of AT commands. E.g., presuming that
SIM-related operations do not depend on QCFG attributes, the new
ordering allows for the execution of SIM-related commands before the
QCFG is fully synchronized.

Fixes #7