modem_pinephone.run: make run script configurable
This patch adds hooks to easily test different aspects of the modem.

First, one can choose between interactively driving the AT command
interface or letting the modem driver speak the AT protocol
automatically. In the latter case, one needs to supply the PIN of the
SIM card as 'SIM_PIN' argument to the run script.

Second, the test for mobile-data connectivity can now be enabled or
disabled.

Third, when operating the AT prococol via the driver, a phone call
can be initiated to the number specified as argument CALL_NUMBER.

By default, the run script produces the simplest scenario possible,
merely offering in interactive console to the modem's AT interface.