sculpt: use left border as virtual KEY_DASHBOARD
This patch implements the toggling between the leitzentrale subsystem
and the runtime subsystem as response to touching the left border
of the touch screen.

The current leitzentrale state is driven by the low-level KEY_DASHBOARD
event generate by the event filter. This state is observed by the phone
manager, which responds to state changes by altering the window layout
of the leitzentrale's window manager. Whenever the runtime is selected
by the user, the menu view of the leitzentrale is moved out of the way.