phone_manager: enable deploy support
This patch adds the sculpt-manager functionality needed to use Sculpt's
package-deploy mechanism. The component graph is hosted in the
"Software" section of the user interface.

The patch renames the dialog classes for the five main sections of the
used interface to avoid name collisions with the types used by the
original sculpt manager (e.g., Storage_dialog, Network_dialog).