run/image/pine_uboot_sdcard: use ext2 defaults
The ext2 parameters given in the pine_uboot_sdcard run-tool plugin
result in a file system suited for small disks, which is fine for
regular Genode scenarios where the boot image is in the order of a few
tens of MiB. However, with Sculpt OS, the ext2 partition is expanded to
the size of the SD-card, e.g., 32 GiB, where those parameters stand in
the way.

In particular, the resizing of the file system from the small profile to
the regular default profile takes minutes. By picking the default
profile as starting point, the resize2fs step works almost instantly.

Fixes genodelabs/genode#4736