at_protocol: allow early modem shutdown
This patch takes the following precautions to accommodate a modem
shutdown triggered early during the modem startup.

First, it retries a failed AT+QPOWD command. The power-down command
can apparently produce an ERROR when issued too early during the modem
startup. This patch re-issues the command in this case.

Second, it gives the AT+QPOWD command precedence over SIM-related
commands.

Third, it reflects a scheduled AT+QPOWD command as 'powering_down'
even if the command is not yet submitted. This can happen during
the early probing phase where the AT commands are issued until the
modem responds with OK. This change results in a more timely
reflection of the state at the user interface.