platform/a64: mitigate panel flickering
The wakeup from the screensaver magnifies the sporadic flickering of the
PinePhone's panel. Apparently, the effect can be mitigated by keeping
the panel powered (at a lowered voltage) while sleeping. An initial
flicker effect when cold-booting the device can still occur but it
usually vanishes after a minute.

Thanks to Josef Soentgen for the investigation of the issue and for the
mitigation approach.