platform: reset SCP-idle only after response
The SCP irq handler must not unconditionally reset the tracked
state to idle but only when observing the completion of a request.

Otherwise, a spurious interrupt can lead to the scheduling of a new
request while the previous request is still executed by the SCP. This
could be observed during the startup of the PinePhone: The platform
driver enabled the lcd0 voltage (via the Local_connection) while the
power driver extended the SCP program (via an Scp::Connection).
In this situation, the command issued by the platform driver could
overwrite the command of the power driver inside SCP memory, usually
resulting in SCP error ("unknown word") while the SCP firmware remained
in compile mode.

Issue #29