a64_wifi: supplement functions in slab_common.c
Normally this functions are provided by the kernel when the SLUB
is enabled and used ('mm/slab_common.c' is referenced in the
'source.list' file).

In our case we rely on 'shadow/mm/slab_common.c' and that would
be the natural location to place these functions. However, since
other driver incorporate that compilation unit as well and already
supplement them (either implemented directly or via generated
dummies), provide them locally for the time being to prevent
friction.

Issue genodelabs/genode#5356.